fix lint errors

+++ x-pack/auditbeat/processors/sessionmd/procfs/procfs.go
@@ -20,11 +20,11 @@ import (
 )
 
 func MajorTTY(ttyNr uint32) uint32 {
-	return uint32((ttyNr >> 8) & 0xff)
+	return (ttyNr >> 8) & 0xff
 }
 
 func MinorTTY(ttyNr uint32) uint32 {
-	return uint32(((ttyNr >> 12) & 0xfff00) | (ttyNr & 0xff))
+	return ((ttyNr >> 12) & 0xfff00) | (ttyNr & 0xff)
 }
 
 // this interface exists so that we can inject a mock procfs reader for deterministic testing
